+++ database/migrations/2022_10_18_123232_create_products_table.php
@@ -9,11 +9,6 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up(): void
     {
         Schema::create('products', function (Blueprint $table) {
@@ -33,14 +28,11 @@ return new class extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
+    public function down(): void
     {
-        Schema::dropIfExists('category_product');
-        Schema::dropIfExists('products');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('category_product');
+            Schema::dropIfExists('products');
+        }
     }
 };
